1_projects/phoenix_mamba_v2/pyproject.toml
[project]
name = "phoenix-mamba-v2"
version = "2.0.0"
description = "Next-Generation Brain Tumor Detection with Selective State-Space Models"
readme = "docs/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Phoenix Protocol Team"}
]
dependencies = [
    # Deep Learning Framework
    "tensorflow>=2.11.0",
    "keras>=2.10.0",

    # Data Processing
    "numpy>=1.23.0",
    "pandas>=1.5.0",
    "opencv-python>=4.6.0",
    "Pillow>=9.2.0",

    # Visualization
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",

    # Machine Learning Utilities
    "scikit-learn>=1.1.0",
    "optuna>=3.1.0",
    "tensorflow-model-optimization>=0.7.0",

    # Progress & Logging
    "tqdm>=4.64.0",
    "mlflow>=2.2.0",

    # Image Augmentation & Analysis
    "imgaug>=0.4.0",
    "scipy>=1.9.0",
    "imagehash>=4.3.0",
    "scikit-image>=0.20.0",

    # Model Export & Serving
    "h5py>=3.7.0",
    "tf2onnx>=1.14.0",
    "onnx>=1.13.0",
    "onnxruntime>=1.14.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",

    # Utilities
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.260"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
